refactor(Add-tables/job_extensions-and-tables/places-folder): Add beter compartimentation of the different files
