+++ sql/tables/__init__.py
@@ -1,6 +1,6 @@
-from .distances import Distances
+from sql.tables.places.distances import Distances
 from .jobs import Jobs
-from .keywords import Keywords
-from .metadata import Metadata
-from .time_stamps import TimeStamps
+from sql.tables.job_extensions.keywords import Keywords
+from sql.tables.job_extensions.metadata import Metadata
+from sql.tables.job_extensions.time_stamps import TimeStamps
 from .base_table import BaseTable
