+++ tests/sql/tables/job_extensions/test_keywords.py
@@ -1,21 +1,20 @@
-from sql.tables.time_stamps import TimeStamps
+from sql.tables.job_extensions.keywords import Keywords
 from sql.tables.jobs import Jobs
 from tests.conftest import BaseTest
 import pytest
-import datetime
 
 @pytest.mark.js_tables
-class TestTimeStamps(BaseTest):
+class TestKeywords(BaseTest):
     """Ensure that Distances table works as intended"""
-    def test_get_job_associated_time_stamps(self):
-        """Test that <get_job_associated_time_stamps> works as intended"""
+    def test_get_job_associated_keywords(self):
+        """Test that <get_job_associated_keywords> works as intended"""
         db1 = f"{self.test_folder}/database.db"
         with Jobs.get_session(db1) as session:
             session.add_all(self.make_small_database())
 
         with Jobs.get_session(db1) as session:
-            assert 2 == len(TimeStamps.get_job_associated_time_stamps(session, "Alpha"))
-            assert 1 == len(TimeStamps.get_job_associated_time_stamps(session, "Beta"))
+            assert 2 == len(Keywords.get_job_associated_keywords(session, "Alpha"))
+            assert 1 == len(Keywords.get_job_associated_keywords(session, "Beta"))
 
     def test_relationship(self):
         """Ensure that the delete cascade works."""
@@ -29,18 +28,18 @@ class TestTimeStamps(BaseTest):
             session.add_all(db)
 
         with Jobs.get_session(db_path) as session:
-            number_of_kw = len(TimeStamps.get_all(session).all())
-            number_of_kw_associated_to_url = len(TimeStamps.get_job_associated_time_stamps(session, url=url))
+            number_of_kw = len(Keywords.get_all(session).all())
+            number_of_kw_associated_to_url = len(Keywords.get_job_associated_keywords(session, url=url))
             session.delete(db[0])
 
         # Now Only
         with Jobs.get_session(db_path) as session:
             # Ensure that passive_deletes works !
-            assert len(TimeStamps.get_all(session).all()) == number_of_kw - number_of_kw_associated_to_url
+            assert len(Keywords.get_all(session).all()) == number_of_kw - number_of_kw_associated_to_url
 
 
     # Utils
-    def make_small_database(self) -> list[Jobs | TimeStamps]:
+    def make_small_database(self) -> list[Keywords | Jobs]:
         """Make a small database for test purposes."""
         j1 = Jobs(
             url="Alpha"
@@ -49,26 +48,26 @@ class TestTimeStamps(BaseTest):
             url="Beta"
         )
 
-        t1 = TimeStamps(
+        k1 = Keywords(
             url="Alpha",
-            label="Download",
-            time_stamp=datetime.datetime.now(),
+            keyword="Biology",
+            occurrence=45,
         )
-        t2 = TimeStamps(
+        k2 = Keywords(
             url="Alpha",
-            label="Informatic",
-            time_stamp=datetime.datetime.now(),
+            keyword="Informatic",
+            occurrence=5,
         )
-        t3 = TimeStamps(
+        k3 = Keywords(
             url="Beta",
-            label="Download",
-            time_stamp=datetime.datetime.now(),
+            keyword="Biology",
+            occurrence=None,
         )
 
         self.screen_multiple_vars("j", j1, j2)
-        self.screen_multiple_vars("t", t1, t2, t3)
+        self.screen_multiple_vars("k", k1, k2, k3)
 
-        return [j1, j2, t1, t2, t3]
+        return [j1, j2, k1, k2, k3]
